backend/requirements.txt
fastapi
uvicorn
python-dotenv
torch
transformers
bitsandbytes
accelerate
pillow
sentinelhub
shapely
pandas
numpy
xgboost
scikit-learn
optuna
rasterio
boto3
opencv-python
python-multipart
